All One Stage Simulation in Action

Conducting one stage of election.

oneset(alike_threshold = 0.2, segregation_radius = 1,
  signalbias_mean = 0.2, signalbias_sd = 0.2, signalbias_size = NULL,
  signalbias_ingroup = 0.5, signaldifficulty = 3,
  updatecommonfactor_sd = 0.1, updatecommonfactor_mean = 0,
  updategroupfactor_sd = 0, updategroupfactor_mean = 0,
  updatepersonalfactor_sd = 0, updatepersonalfactor_mean = 0,
  partyloc_noise = 0, partycapacity_noise = 0.5,
  partyblue_loc_direction = 0, partyred_loc_direction = 0,
  partycapacity_direction_mean = 0, partyupdate_type = "oneshot",
  context_type = "region", contextradius = 5, contexttimetoagg = 2,
  b_idedist = 0.198, b_capacity = 0.716, b_incumbent = 1.526,
  b_pvi_nation = -9.827, b_pvi_local = 5.156, b_pvi_sublocal = 1.663,
  b_pid = 1.215, maxt_segregation = 50, n_elections = 30,
  n_drop_elections = 5, n_voters = 2300, spacelim = c(54, 54),
  sizegroup = c(0.5, 0.5), regionlevel = c(3, 3),
  subregionlevel = c(9, 9), startcommonfactor = NULL,
  startgroupfactor = 0.8, startpersonalfactor = 1,
  groupfactor_type = "constant", knowmean1 = 0.65, knowmean2 = 0.65,
  knowspread = 0.85, partyblue_startloc = -0.8,
  partyred_startloc = 0.8, partystartloc_noise = 0,
  partyblue_startcapacity = 0, partyred_startcapacity = 0,
  partystartcapacity_noise = 0.5, burnin_redvoteprob_sd = 0.05,
  burnin_redvoteprob_cor = 0.19, seedval = NULL, drawplot = FALSE,
  run_type = "all", segregated_soc = NULL)

Arguments

alike_threshold
1 - tolerance in the segregation game. The default = 0.2;

segregation_radius
Search radius to look for neightbors The default = 1; 8 people around

signalbias_mean
The average common directional bias (i.e., positive value advantages red party and vice versa) in signal. The default is 0.5.

signalbias_sd
The SD common directional bias. The bias is drawn from normal distribution each time. The default is 0.5.

signalbias_size
The extent of fixed common directional bias. If assigned; mean and sd arguments above will be ignored.

signalbias_ingroup
The extent of fixed common ingroup bias (i.e., Evaluate the ingroup candidate better). The default is 0.

signaldifficulty
The extent of noise level in signal. Given the default values of knowledge, ideology, and party position, the default difficulty of 3 produces idedist with SD 1.9; Consistent with ANES.

updatecommonfactor_sd
The movement in common factor. SD. The default is 0.1.

updatecommonfactor_mean
The movement in common factor. Mean. The default is 0.

updategroupfactor_sd
The movement in group influence on ideology. SD. The default is 0.

updategroupfactor_mean
The movement in group influence on ideology. Mean. The default is 0.

updatepersonalfactor_sd
The SD for the movemebnt in personal influence on ideology. SD. The default is 0.

updatepersonalfactor_mean
The SD for the movemebnt in personal influence on ideology. Mean. The default is 0.

partyloc_noise
Random noise in updating locations of party ideologies. The default is 0.

partycapacity_noise
Random noise in updating party capacity. The default is 0.5. The average score in ANES is 0.7 but it produces too much movement in voting.

partyblue_loc_direction
Direction and expected quantity of the movement in blue party ideology.

partyred_loc_direction
Direction and expected quantity of the movement in red party ideology.

partycapacity_direction_mean
The average direction and expected quantity of the movement in party capacity. (Positive vaue advantages red party and negative value advantages blue party).

partyupdate_type
The type of updating procedure. If "oneshot", the update is always given over startloc and startcapacity defined in setparty at each election. If "online", the update is given over the ideology and capacity realized in the previous election. If "fromstart" the update is given over the ideology and capacity realized in the first election of the simulation run.

context_type
"region" (default) or "network" (currently not used).

contextradius
Radius to seach neighbors. Only applied when context_type=="network" (currently not used).

contexttimetoagg
How many previous elections to average. Either 1 or 2 (default).

b_idedist
The influence of ideological distance. The default = 0.198, from ANES.

b_capacity
The influence of capacity The default = 0.716, from ANES.

b_incumbent
The influence if incumebent. The default = 1.526, from ANES (EDV result).

b_pvi_nation
The influence of national PVI. The default = -9.827, from ANES (EDV result).

b_pvi_local
The influence of regional PVI. The default = 5.156, from CCES pvi_state.

b_pvi_sublocal
The influence of sub-regional PVI. The default = 1.663, from CCES pvi_state.

b_pid
The influence of group membership. The default = 1.215, from ANES.

maxt_segregation
Max time of movement in the segregation game. The default is 50.

n_elections
Number of elections in each simulation run. The default is 30.

n_drop_elections
Number of first N elections dropped from the quality assessment. The default is 5.

n_voters
Interger. Number of agents. The default is 2300.

spacelim
Integer vector of length 2. Grid space size (number of rows and number of columns, respectively). The default is c(54,54).

sizegroup
Numeric Vector. Proportion of agents in each group. Length defines the number of groups. Must adds up to 1. The default is c(0.5,0.5).

regionlevel
Integer vector of length 2. Level to Split Region (Horizontaly & Vertically). The default is c(3,3).

subregionlevel
Ingeger vector of length 2. Level to Split Sub-Region (Horizontaly & Vertically). The default is c(9,9).

startcommonfactor
The starting value of The common factor in ideology (intercept). The default is NULL (draw randomly).

startgroupfactor
The starting value of the group influence on ideology. The default is 0.8 (from ANES, skewed slightly towards right in the reality).

startpersonalfactor
The starting value of the random personal factor in ideology (The SD of the random error to the model). The defauls is 1 (to get ideology distribution with SD 1, from ANES).

groupfactor_type
If "constant", groupfactor is the constant coefficient for group. If "diverging", groupfactor is the SD for folded normal distribution that randomly assigns coeffcient for group.

knowmean1
The lowest average level of knowledge for Group Blue/-1. 0.65 is the overall mean in ANES (stable).

knowmean2
The highest average level of knowledge for Group Red/1. 0.65 is the overall mean in ANES (stable).

knowspread
The spread (SD for normal distribution) of knowledg within a group. The default is 0.85 (To get approx 0.25 SD after conversion, referred to ANES).

partyblue_startloc
Starting location of blue party ideology. The default is -0.8 (from ANES).

partyred_startloc
Starting location of red party ideology. The default is 0.8 (from ANES).

partystartloc_noise
Random noise given to the starting locations of party ideologies. The default is 0.

partyblue_startcapacity
Starting level of blue party capacity. The default is 0.

partyred_startcapacity
Starting level of red party capacity. The default is 0.

partystartcapacity_noise
Random noise given to the starting level of party capacities. The default is 0.5.

burnin_redvoteprob_sd
The spread for the "burn-in" period (i.e., first two elections) voting behaviors. The default is 0.05, referred to SD in ANES National PVI. See details.

burnin_redvoteprob_cor
The correlation between voting and group in "burn-in" period voting behaviors. the default is 0.19. referred to ANES. See details.

seedval
The random number seed.

drawplot
If True; draw plots of results

run_type
The default is "all" (to run both segregation and election models) Can also be "segregation" or "election" (which runs only one of them).

segregated_soc
The pre-defined society object. If not NULL and run_type == "election", predefined society is used in simulation.

Value

List object with society characteristics, including following elements:

grid The grid of voters (with group membership at each cell)

grid_region The grid of region identifier at each cell.

grid_subregion The grid of sub-region identifier at each cell.

grid_coords All grid coordinates.

group The numeric vector of group memberships

region_stat Aggregated group statistics by region.

subregion_stat Aggregated group statistics by region.

grid_votes2 The grid of voting decisions in t-2 election.

grid_votes1 The grid of voting decisions in t-1 election.

ideology Created by setideology, may be updated by updateideology

knowledge Created by setknowledge

context Created by getregion

signal Created by getsignal

incumbent Character of incumbent party. Either "blue" or "red".

votes Created by letsvote

quality Aggregated quality of decisions.

Details

In the first two elections, the model does not have the sufficient number of previous elections to aggregate. If the society does not cotain the information of two previous elections, the hypothetical votes in first two election is radomly provided by the following procedure.

Draw the reference value of voting for red for each group of voters by multi-variate normal distribution with mean 0 and the covariance with group membership (i.e., -1, or 1) set by burnin_redvoteprob_cor.

Multiply the reference values by burnin_redvoteprob_sd and add 0.5 to generate the probability of red vote for each group.

Draw vote of voters in each group by the Bernouli distribution with probability drawin in the previous step.