All One Stage Voting in Action

Conducting one stage of election.

prefsoc(soc, commonfactor = 0.1, groupfactor = 0.8,
  personalfactor = 1, updatecommonfactor = 0.1,
  updategroupfactor = 0, updatepersonalfactor = 0,
  groupfactor_type = "constant", knowmean1 = 0.65, knowmean2 = 0.65,
  knowspread = 0.85, partyblue_startloc = -0.8,
  partyred_startloc = 0.8, partystartloc_noise = 0.5,
  partyblue_startcapacity = 0, partyred_startcapacity = 0,
  partystartcapacity_noise = 0.5, partyloc_noise = 0,
  partycapacity_noise = 0, partyblue_loc_direction = 0,
  partyred_loc_direction = 0, partyblue_capacity_direction = 0,
  partyred_capacity_direction = 0, partyupdate_type = "oneshot",
  signaldifficulty = 3, signalbias_common = 0,
  signalbias_ingroup = 0, context_type = "region", contextradius = 5,
  contexttimetoagg = 2, b_idedist = 0.198, b_capacity = 0.716,
  b_incumbent = 1.526, b_pvi_nation = -9.827, b_pvi_local = 5.156,
  b_pvi_sublocal = 1.663, b_pid = 1.215,
  burnin_redvoteprob_sd = 0.045, burnin_redvoteprob_cor = 0.19,
  seedval = NULL)

Arguments

soc
The lists generated initially by initialsoc.

commonfactor
The common factor in ideology (intercept).

groupfactor
The group influence on ideology.

personalfactor
The random personal factor in ideology (The SD of the random error to the model).

updatecommonfactor
The movement in common factor.

updategroupfactor
The movement in group influence on ideology.

updatepersonalfactor
The SD for the movemebnt in personal influence on ideology.

groupfactor_type
If "constant", groupfactor is the constant coefficient for group. If "diverging", groupfactor is the SD for folded normal distribution that randomly assigns coeffcient for group.

knowmean1
The lowest average level of knowledge for a group.

knowmean2
The highest average level of knowledge for a group.

knowspread
The spread (SD for normal distribution) of knowledg within a group.

partyblue_startloc
Starting location of blue party ideology.

partyred_startloc
Starting location of red party ideology.

partystartloc_noise
Random noise given to the starting locations of party ideologies.

partyblue_startcapacity
Starting level of blue party capacity.

partyred_startcapacity
Starting level of red party capacity.

partystartcapacity_noise
Random noise given to the starting level of party capacities.

partyloc_noise
Random noise in updating locations of party ideologies.

partycapacity_noise
Random noise in updating party capacity.

partyblue_loc_direction
Direction and expected quantity of the movement in blue party ideology.

partyred_loc_direction
Direction and expected quantity of the movement in red party ideology.

partyblue_capacity_direction
Direction and expected quantity of the movement in blue party capacity.

partyred_capacity_direction
Direction and expected quantity of the movement in red party capacity.

partyupdate_type
The type of updating procedure. If "oneshot", the update is always given over startloc and startcapacity defined in setparty at each election. If "online", the update is given over the ideology and capacity realized in the previous election. If "fromstart" the update is given over the ideology and capacity realized in the first election of the simulation run.

signaldifficulty
The extent of noise level in signal.

signalbias_common
The extent of common directional bias (i.e., positive value advantages red party and vice versa) in signal.

signalbias_ingroup
The extent of common ingroup bias (i.e., Evaluate the ingroup candidate better).

context_type
"region" (default) or "network" (currently not used).

contextradius
Radius to seach neighbors. Only applied when context_type=="network" (currently not used).

contexttimetoagg
How many previous elections to average. Either 1 or 2 (default).

b_idedist
The influence of ideological distance. The default = 0.198, # From ANES

b_capacity
The influence of capacity The default = 0.716, # From ANES

b_incumbent
The influence if incumebent. The default = 1.526, # From ANES (EDV result)

b_pvi_nation
The influence of national PVI. The default = -9.827, # From ANES (EDV result)

b_pvi_local
The influence of regional PVI. The default = 5.156, # From CCES pvi_state

b_pvi_sublocal
The influence of sub-regional PVI. The default = 1.663, # From CCES pvi_state

b_pid
The influence of group membership. The default = 1.215, # From ANES

burnin_redvoteprob_sd
The spread for the "burn-in" period (i.e., first two elections) voting behaviors. See details.

burnin_redvoteprob_cor
The correlation between voting and group in "burn-in" period voting behaviors. See details.

seedval
The random number seed.

Value

List object with society characteristics, including following elements:

grid The grid of voters (with group membership at each cell)

grid_region The grid of region identifier at each cell.

grid_subregion The grid of sub-region identifier at each cell.

grid_coords All grid coordinates.

group The numeric vector of group memberships

region_stat Aggregated group statistics by region.

subregion_stat Aggregated group statistics by region.

grid_votes2 The grid of voting decisions in t-2 election.

grid_votes1 The grid of voting decisions in t-1 election.

ideology Created by setideology, may be updated by updateideology

knowledge Created by setknowledge

context Created by getregion

signal Created by getsignal

incumbent Character of incumbent party. Either "blue" or "red".

votes Created by letsvote

Details

In the first two elections, the model does not have the sufficient number of previous elections to aggregate. If the society does not cotain the information of two previous elections, the hypothetical votes in first two election is radomly provided by the following procedure.

Draw the reference value of voting for red for each group of voters by multi-variate normal distribution with mean 0 and the covariance with group membership (i.e., -1, or 1) set by burnin_redvoteprob_cor.

Multiply the reference values by burnin_redvoteprob_sd and add 0.5 to generate the probability of red vote for each group.

Draw vote of voters in each group by the Bernouli distribution with probability drawin in the previous step.