Set Characteristics of Two Parties

Set ideologies and capacity of parties (i.e., blue and red) running in election.

setparty(blue_startloc = -0.8, red_startloc = 0.8,
  startloc_noise = 0.1, blue_startcapacity = 0,
  red_startcapacity = 0, startcapacity_noise = 0.1,
  startincumbent = NULL, blue_loc_seed = NULL, red_loc_seed = NULL,
  blue_capacity_seed = NULL, red_capacity_seed = NULL)

Arguments

blue_startloc
Starting location of blue party ideology.

red_startloc
Starting location of red party ideology.

startloc_noise
Random noise given to the starting locations of party ideologies.

blue_startcapacity
Starting level of blue party capacity.

red_startcapacity
Starting level of red party capacity.

startcapacity_noise
Random noise given to the starting level of party capacities.

startincumbent
Incumbent in the first election. If NULL, randomly chosen.

blue_loc_seed
Random number seed to draw blue party ideology.

red_loc_seed
Random number seed to draw red party ideology.

blue_capacity_seed
Random number seed to draw blue party capacity.

red_capacity_seed
Random number seed to draw red party capacity.

Value

The updated list of ideology parameters

loc ideology locations of two parties.

capacity capacity levels of two parties.

tracker Tracker of the history of ideogoloy and capacity.